Address the review comments: - Move "JobDefinition" to internal package - Deprecate toDag() method.
